+++ Algorithms/Dft.cs
@@ -9,9 +9,9 @@ namespace Algorithms
 	public class DftParameters : AlgorithmParameters
 	{
 		public readonly int Key;
-		public readonly double Alpha;
+		public readonly decimal Alpha;
 
-		public DftParameters(EffectiveBitmap original, EffectiveBitmap watermark, EffectiveBitmap watermarked, int key, double alpha) : base(original, watermark, watermarked)
+		public DftParameters(EffectiveBitmap original, EffectiveBitmap watermark, EffectiveBitmap watermarked, int key, decimal alpha) : base(original, watermark, watermarked)
 		{
 			Key = key;
 			Alpha = alpha;
@@ -52,7 +52,7 @@ namespace Algorithms
 			for (int i = 0; i < complexWatermarked.Width; i++)
 			{
 				v[i] = random.Next(0, 2);
-				vAlpha[i] = v[i] * parameters.Alpha;
+				vAlpha[i] = v[i] * (double)parameters.Alpha;
 			}
 
 			for (int y = 0; y < complexWatermarked.Height; y++)
@@ -78,7 +78,7 @@ namespace Algorithms
 			for (int i = 0; i < complexWatermarked.Width; i++)
 			{
 				v[i] = random.Next(0, 2);
-				vAlpha[i] = v[i] * parameters.Alpha;
+				vAlpha[i] = v[i] * (double)parameters.Alpha;
 			}
 
 			for (int y = 0; y < complexWatermarked.Height; y++)
@@ -101,7 +101,7 @@ namespace Algorithms
 			for (int i = 0; i < complexImage.Width; i++)
             {
 				v[i] = random.Next(0, 2);
-				vAlpha[i] = v[i] * parameters.Alpha;
+				vAlpha[i] = v[i] * (double)parameters.Alpha;
 			}
 				
 			for (int y = 0; y < complexImage.Height; y++)
